Renamed FramesTester and removed unused imports
