+++ tests/test_graphical_units/test_creation.py
@@ -1,7 +1,6 @@
 import pytest
 
 from manim import *
-from tests.test_graphical_units.testing.frames_comparison import frames_comparison
 
 from ..utils.GraphicalUnitTester import GraphicalUnitTester
 from ..utils.testing_utils import get_scenes_to_test
